Run time data

Run times measured on an Intel(R) Xeon(R) CPU X5680 clocked at 3.33 GHz with 12 MB cache, running RedHat Enterprise Linux, R 3.2.3 and gcc 4.8.5, using version 0.4 of the wrswoR package. The data are created by the corresponding scripts in the data_raw directory.

timings_sort

Format

An object of class data.frame with 25200 rows and 5 columns.

Details

timings contains run times for a larger range of values for the n argument.

timings_sort contains run times for sorting probabilities with the given distributions.

break_even contains detailed run times for the analysis of break-even points between the various implementations.

Examples

head(timings)

#>      prob mix  expr    time    r  n
#> 1 uniform asc crank 1596193 0.01 31
#> 2 uniform asc crank    8557 0.01 31
#> 3 uniform asc crank    6263 0.01 31
#> 4 uniform asc crank    6159 0.01 31
#> 5 uniform asc crank    5672 0.01 31
#> 6 uniform asc crank    5900 0.01 31

head(break_even)

#>      prob mix  expr    time    r  n
#> 1 uniform asc crank 8479190 0.01 10
#> 2 uniform asc crank    9265 0.01 10
#> 3 uniform asc crank    9406 0.01 10
#> 4 uniform asc crank    5057 0.01 10
#> 5 uniform asc crank    9920 0.01 10
#> 6 uniform asc crank    5899 0.01 10